rename examples
