--- examples/gatsby-sourcing-wordpress/gatsby-node.js
@@ -1,60 +1,0 @@
-const path = require(`path`)
-const slash = require(`slash`)
-
-exports.createPages = async ({ graphql, actions }) => {
-  const { createPage } = actions
-
-  const result = await graphql(`
-    {
-      allWordpressPage {
-        edges {
-          node {
-            id
-            slug
-            status
-            template
-          }
-        }
-      }
-      allWordpressPost {
-        edges {
-          node {
-            id
-            slug
-            status
-            template
-            format
-          }
-        }
-      }
-    }
-  `)
-
-  if (result.errors) {
-    console.error(result.errors)
-  }
-
-  const { allWordpressPage, allWordpressPost } = result.data
-
-  const pageTemplate = path.resolve(`./src/templates/page.js`)
-  allWordpressPage.edges.forEach(edge => {
-    createPage({
-      path: `/${edge.node.slug}/`,
-      component: slash(pageTemplate),
-      context: {
-        id: edge.node.id,
-      },
-    })
-  })
-
-  const postTemplate = path.resolve(`./src/templates/post.js`)
-  allWordpressPost.edges.forEach(edge => {
-    createPage({
-      path: `/${edge.node.slug}/`,
-      component: slash(postTemplate),
-      context: {
-        id: edge.node.id,
-      },
-    })
-  })
-}
