--- examples/gatsby-sourcing-wordpress/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "gatsby-example-sourcing-wordpress",
-  "private": true,
-  "description": "An example to learn how to source data form Wordpress",
-  "version": "0.1.0",
-  "author": "Vishwa Mehta <vishwa.mehta30@gmail.com>",
-  "dependencies": {
-    "gatsby": "^2.13.73",
-    "gatsby-source-wordpress": "^3.0.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-helmet": "^5.2.1"
-  },
-  "devDependencies": {
-    "prettier": "^1.18.2"
-  },
-  "keywords": [
-    "gatsby"
-  ],
-  "license": "MIT",
-  "scripts": {
-    "build": "gatsby build",
-    "develop": "gatsby develop",
-    "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
-    "start": "npm run develop",
-    "serve": "gatsby serve"
-  }
-}
