--- examples/gatsby-sourcing-wordpress/src/pages/index.js
@@ -1,58 +1,0 @@
-import React, { Component } from "react"
-import { Link, graphql } from "gatsby"
-
-class Homepage extends Component {
-  render() {
-    const data = this.props.data
-
-    return (
-        <div>
-          <h1>Pages</h1>
-          {data.allWordpressPage.edges.map(({ node }) => (
-            <div key={node.slug}>
-              <Link to={node.slug}>
-                <h2>{node.title}</h2>
-              </Link>
-              <h3>{node.excerpt}</h3>
-            </div>
-          ))}
-        </div>
-
-        <h1>Posts</h1>
-        {data.allWordpressPost.edges.map(({ node }) => (
-          <div key={node.slug}>
-            <Link to={node.slug}>
-              <h2>{node.title}</h2>
-            </Link>
-            <h3>{node.excerpt}</h3>
-          </div>
-        ))}
-    )
-  }
-}
-
-export default Homepage
-
-export const pageQuery = graphql`
-  query {
-    allWordpressPage {
-      edges {
-        node {
-          id
-          title
-          excerpt
-          slug
-        }
-      }
-    }
-    allWordpressPost {
-      edges {
-        node {
-          title
-          excerpt
-          slug
-        }
-      }
-    }
-  }
-`
