--- examples/gatsby-sourcing-wordpress/src/templates/page.js
@@ -1,33 +1,0 @@
-import React, { Component } from "react"
-import { graphql } from "gatsby"
-
-class Page extends Component {
-  render() {
-    const StaticPage = this.props.data.wordpressPage
-
-    return (
-        <>
-        <h1>{StaticPage.title }</h1>
-        <div>{StaticPage.content}</div/>
-        </>
-    )
-  }
-}
-
-export default Page
-
-export const pageQuery = graphql`
-  query($id: String!) {
-    wordpressPage(id: { eq: $id }) {
-      title
-      content
-    }
-    site {
-      id
-      siteMetadata {
-        title
-        subtitle
-      }
-    }
-  }
-`
