--- examples/gatsby-sourcing-wordpress/src/templates/post.js
@@ -1,37 +1,0 @@
-import React, { Component } from "react"
-import { graphql } from "gatsby"
-import PropTypes from "prop-types"
-
-class Post extends Component {
-  render() {
-    const post = this.props.data.wordpressPost
-
-    return (
-      <>
-        <h1>{post.title}</h1>
-        <div>{post.content}</div>
-      </>
-    )
-  }
-}
-
-Post.propTypes = {
-  data: PropTypes.object.isRequired,
-  edges: PropTypes.array,
-}
-
-export default Post
-
-export const pageQuery = graphql`
-  query($id: String!) {
-    wordpressPost(id: { eq: $id }) {
-      title
-      content
-    site {
-      siteMetadata {
-        title
-        subtitle
-      }
-    }
-  }
-`
