+++ examples/using-wordpress-with-acf/README.md
@@ -1,8 +1,8 @@
-# Using Wordpress
+# Using Wordpress With ACF
 
 https://using-wordpress.gatsbyjs.org/
 
-Example site that demonstrates how to build Gatsby sites that pull data from the
-[WordPress CMS API](https://www.wordpress.com/).
+Example site that extends the [using-wordpress](examples/using-wordpress/) example and demonstrates how to build Gatsby sites that pull data from the
+[WordPress CMS API](https://www.wordpress.com/) using ACF.
 
 IMPORTANT NOTE: This example site won't keep working if you switch the WordPress site to one of yours. It's meant as a learning resource. Try starting from scratch with a site created by `gatsby new` and start copying over parts from this site starting with `gatsby-config.js`.
